[CASSH Server] v1.8.0: Python3 support
