+++ src/server/Dockerfile
@@ -1,19 +1,20 @@
-FROM debian:latest
+FROM python:3.6
 
-# Container options (For layer optim)
 WORKDIR /opt/cassh
-ENTRYPOINT ["/opt/cassh/server/docker-entrypoint"]
 
 RUN apt-get update \
     && apt-get install -yqq \
         openssh-client \
         openssl \
-        python-psycopg2 \
-        python-webpy \
-        python-ldap \
-        python-configparser \
-        python-requests \
-        python-openssl \
+        libldap2-dev \
+        libsasl2-dev \
     && rm -rf /var/lib/apt/lists/*
 
-COPY . server/
+
+COPY requirements.txt ./
+
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY . .
+
+CMD [ "python", "./server.py" ]
