+++ src/server/conf/cassh.conf
@@ -4,6 +4,7 @@ krl = __CASSH_PATH__/test-keys/revoked-keys
 port = 8080
 cluster = http://localhost:8080,http://localhost:8081
 clustersecret = clustersecretpassword
+debug = False
 
 [postgres]
 host = localhost
