+++ src/server/server.py
@@ -3,30 +3,22 @@
 """
 Sign a user's SSH public key.
 """
-from __future__ import print_function
 from argparse import ArgumentParser
-from datetime import datetime, timedelta
 from json import dumps
-from os import remove, stat
-from random import choice
+from os import remove
 from re import compile as re_compile
-from shutil import move
-from string import ascii_lowercase
 from tempfile import NamedTemporaryFile
-from time import time
-from urllib import unquote_plus
+from urllib.parse import unquote_plus
 
 # Third party library imports
 from configparser import ConfigParser, NoOptionError
-from ldap import open as ldap_open, SCOPE_SUBTREE
-from psycopg2 import connect, OperationalError, ProgrammingError
-from requests import Session
-from requests.exceptions import ConnectionError
-from web import application, config, ctx, data, httpserver
+from ldap import initialize, SCOPE_SUBTREE
+from web import application, config, data, httpserver
 from web.wsgiserver import CherryPyWSGIServer
 
 # Own library
 from ssh_utils import Authority, get_fingerprint
+from tools import get_principals, random_string, response_render, unquote_custom, Tools
 
 # DEBUG
 # from pdb import set_trace as st
@@ -50,13 +42,7 @@ URLS = (
     '/test_auth', 'TestAuth',
 )
 
-VERSION = '1.7.0'
-
-REQ_HEADERS = {
-    'User-Agent': 'CASSH-SERVER v%s' % VERSION,
-    'SERVER_VERSION': VERSION,
-}
-REQ_TIMEOUT = 2
+VERSION = '1.8.0'
 
 PARSER = ArgumentParser()
 PARSER.add_argument('-c', '--config', action='store', help='Configuration file')
@@ -127,209 +113,27 @@ try:
     SERVER_OPTS['clustersecret'] = CONFIG.get('main', 'clustersecret')
 except NoOptionError:
     # Standalone mode
-    SERVER_OPTS['clustersecret'] = randomString(32)
+    SERVER_OPTS['clustersecret'] = random_string(32)
 
+try:
+    SERVER_OPTS['debug'] = bool(CONFIG.get('main', 'debug') != 'False')
+except NoOptionError:
+    SERVER_OPTS['debug'] = False
 
-# INDEPENDANT FUNCTIONS
-def get(url):
-    """
-    Rebuilt GET function for CASSH purpose
-    """
-    session = Session()
-    try:
-        req = session.get(url,
-                          headers=REQ_HEADERS,
-                          timeout=REQ_TIMEOUT,
-                          stream=True)
-    except ConnectionError:
-        print('Connection error : %s' % url)
-        req = None
-    return req
-
-def post(url, payload):
-    """
-    Rebuilt POST function for CASSH purpose
-    """
-    session = Session()
-    try:
-        req = session.post(url,
-                           data=payload,
-                           headers=REQ_HEADERS,
-                           timeout=REQ_TIMEOUT)
-    except ConnectionError:
-        print('Connection error : %s' % url)
-        req = None
-    return req
-
-def randomString(stringLength=10):
-    """Generate a random string of fixed length """
-    letters = ascii_lowercase
-    return ''.join(choice(letters) for i in range(stringLength))
-
-def cluster_alived():
-    """
-    This function returns a subset of pingeable node
-    """
-    alive_nodes = list()
-    dead_nodes = list()
-    for node in SERVER_OPTS['cluster']:
-        req = get("%s/ping" % node)
-        if req is not None and req.text == 'pong':
-            alive_nodes.append(node)
-        else:
-            dead_nodes.append(node)
-    return alive_nodes, dead_nodes
-
-def cluster_last_krl():
-    """
-    This functions download the biggest KRL of the cluster.
-    It's not perfect...
-    """
-    subset, _ = cluster_alived()
-    cluster_id = 0
-    for node in subset:
-        req = get("%s/krl" % node)
-        with open('/tmp/%s.krl' % cluster_id, 'wb') as cluster_file:
-            for chunk in req.iter_content(chunk_size=1024):
-                if chunk: # filter out keep-alive new chunks
-                    cluster_file.write(chunk)
-        cluster_id += 1
-
-    top_size = stat(SERVER_OPTS['krl']).st_size
-    top_cluster_id = 'local'
-    for i in range(cluster_id):
-        if stat('/tmp/%s.krl' % i).st_size >= top_size:
-            top_cluster_id = i
-
-    print('%s KRL is the best' % top_cluster_id)
-
-    if top_cluster_id != 'local':
-        move('/tmp/%s.krl' % top_cluster_id, SERVER_OPTS['krl'])
-    else:
-        cluster_updatekrl(None, update_only=True)
-
-    return True
-
-
-def cluster_updatekrl(username, update_only=False):
-    """
-    This function send the revokation of a user to the cluster
-    """
-    subset, _ = cluster_alived()
-    reqs = list()
-    payload = {"clustersecret": SERVER_OPTS['clustersecret']}
-    if not update_only:
-        payload.update({"username": username})
-    for node in subset:
-        req = post("%s/cluster/updatekrl" % node, payload)
-        reqs.append(req)
-    return reqs
+TOOLS = Tools(SERVER_OPTS, STATES, VERSION)
 
 def data2map():
     """
     Returns a map from data POST
     """
     data_map = {}
-    if data() == '':
+    data_str = data().decode('utf-8')
+    if data_str == '':
         return data_map
-    for key in data().split('&'):
+    for key in data_str.split('&'):
         data_map[key.split('=')[0]] = '='.join(key.split('=')[1:])
     return data_map
 
-def get_principals(sql_result, username, shell=False):
-    """
-    Transform sql principals into readable one
-    """
-    if sql_result is None or sql_result == '':
-        if shell:
-            return username
-        return [username]
-    else:
-        if shell:
-            return sql_result
-        return sql_result.split(',')
-
-def unquote_custom(string):
-    """
-    Returns True is the string is quoted
-    """
-    if ' ' not in string:
-        string = unquote_plus(string)
-        if '+' in string and '%20' in string:
-            # Old custom quote
-            string = string.replace('%20', ' ')
-    return string
-
-def pg_connection(
-        dbname=SERVER_OPTS['db_name'],
-        user=SERVER_OPTS['db_user'],
-        host=SERVER_OPTS['db_host'],
-        password=SERVER_OPTS['db_password']):
-    """
-    Return a connection to the db.
-    """
-    message = ''
-    try:
-        pg_conn = connect("dbname='%s' user='%s' host='%s' password='%s'"\
-            % (dbname, user, host, password))
-    except OperationalError:
-        return None, 'Error : Server cannot connect to database'
-    try:
-        pg_conn.cursor().execute("""SELECT * FROM USERS""")
-    except ProgrammingError:
-        return None, 'Error : Server cannot connect to table in database'
-    return pg_conn, message
-
-def pretty_ssh_key_hash(pubkey_fingerprint):
-    """
-    Returns a pretty json from raw pubkey
-    KEY_BITS KEY_HASH [JERK] (AUTH_TYPE)
-    """
-    try:
-        key_bits = int(pubkey_fingerprint.split(' ')[0])
-    except ValueError:
-        key_bits = 0
-    except IndexError:
-        key_bits = 0
-
-    try:
-        key_hash = pubkey_fingerprint.split(' ')[1]
-    except IndexError:
-        key_hash = pubkey_fingerprint
-
-    try:
-        auth_type = pubkey_fingerprint.split('(')[-1].split(')')[0]
-    except IndexError:
-        auth_type = 'Unknown'
-
-    rate = 'UNKNOWN'
-    if auth_type == 'DSA':
-        rate = 'VERY LOW'
-    elif (auth_type == 'RSA' and key_bits >= 4096) or (auth_type == 'ECDSA' and key_bits >= 256):
-        rate = 'HIGH'
-    elif auth_type == 'RSA' and key_bits >= 2048:
-        rate = 'MEDIUM'
-    elif auth_type == 'RSA' and key_bits < 2048:
-        rate = 'LOW'
-    elif auth_type == 'ED25519' and key_bits >= 256:
-        rate = 'VERY HIGH'
-
-    return {'bits': key_bits, 'hash': key_hash, 'auth_type': auth_type, 'rate': rate}
-
-def str2date(string):
-    """
-    change xd => seconds
-    change xh => seconds
-    """
-    delta = 0
-    if 'd' in string:
-        delta = timedelta(days=int(string.split('d')[0])).total_seconds()
-    elif 'h' in string:
-        delta = timedelta(hours=int(string.split('h')[0])).total_seconds()
-    return delta
-
-
-# OTHER FUNCTION
 def ldap_authentification(admin=False):
     """
     Return True if user is well authentified
@@ -338,17 +142,17 @@ def ldap_authentification(admin=False):
     """
     if SERVER_OPTS['ldap']:
         credentials = data2map()
-        if credentials.has_key('realname'):
+        if 'realname' in credentials:
             realname = unquote_plus(credentials['realname'])
         else:
             return False, 'Error: No realname option given.'
-        if credentials.has_key('password'):
+        if 'password' in credentials:
             password = unquote_plus(credentials['password'])
         else:
             return False, 'Error: No password option given.'
         if password == '':
             return False, 'Error: password is empty.'
-        ldap_conn = ldap_open(SERVER_OPTS['ldap_host'])
+        ldap_conn = initialize("ldap://"+SERVER_OPTS['ldap_host'])
         try:
             ldap_conn.bind_s(realname, password)
         except Exception as e:
@@ -365,78 +169,6 @@ def ldap_authentification(admin=False):
                 return False, 'Error: user %s is not an admin.' % realname
     return True, 'OK'
 
-def list_keys(username=None, realname=None):
-    """
-    Return all keys.
-    """
-    pg_conn, message = pg_connection()
-    if pg_conn is None:
-        ctx.status = '503 Service Unavailable'
-        return message
-    cur = pg_conn.cursor()
-    is_list = False
-
-    if realname is not None:
-        cur.execute('SELECT * FROM USERS WHERE REALNAME=lower((%s))', (realname,))
-        result = cur.fetchone()
-    elif username is not None:
-        cur.execute('SELECT * FROM USERS WHERE NAME=(%s)', (username,))
-        result = cur.fetchone()
-    else:
-        cur.execute('SELECT * FROM USERS')
-        result = cur.fetchall()
-        is_list = True
-    cur.close()
-    pg_conn.close()
-    return sql_to_json(result, is_list=is_list)
-
-def sign_key(tmp_pubkey_name, username, expiry, principals, db_cursor=None):
-    """
-    Sign a key and return cert_contents
-    """
-    # Load SSH CA
-    ca_ssh = Authority(SERVER_OPTS['ca'], SERVER_OPTS['krl'])
-
-    # Sign the key
-    try:
-        cert_contents = ca_ssh.sign_public_user_key(\
-            tmp_pubkey_name, username, expiry, principals)
-        if db_cursor is not None:
-            db_cursor.execute('UPDATE USERS SET STATE=0, EXPIRATION=(%s) WHERE NAME=(%s)', \
-                (time() + str2date(expiry), username))
-    except:
-        cert_contents = 'Error : signing key'
-    return cert_contents
-
-def sql_to_json(result, is_list=False):
-    """
-    This function prettify a sql result into json
-    """
-    if result is None:
-        return None
-    if is_list:
-        d_result = {}
-        for res in result:
-            d_sub_result = {}
-            d_sub_result['username'] = res[0]
-            d_sub_result['realname'] = res[1]
-            d_sub_result['status'] = STATES[res[2]]
-            d_sub_result['expiration'] = datetime.fromtimestamp(res[3]).strftime('%Y-%m-%d %H:%M:%S')
-            d_sub_result['ssh_key_hash'] = pretty_ssh_key_hash(res[4])
-            d_sub_result['expiry'] = res[6]
-            d_sub_result['principals'] = get_principals(res[7], res[0])
-            d_result[res[0]] = d_sub_result
-        return dumps(d_result, indent=4, sort_keys=True)
-    d_result = {}
-    d_result['username'] = result[0]
-    d_result['realname'] = result[1]
-    d_result['status'] = STATES[result[2]]
-    d_result['expiration'] = datetime.fromtimestamp(result[3]).strftime('%Y-%m-%d %H:%M:%S')
-    d_result['ssh_key_hash'] = pretty_ssh_key_hash(result[4])
-    d_result['expiry'] = result[6]
-    d_result['principals'] = get_principals(result[7], result[0])
-    return dumps(d_result, indent=4, sort_keys=True)
-
 class Admin():
     """
     Class admin to action or revoke keys.
@@ -446,7 +178,9 @@ class Admin():
         DEPRECATED Status
         """
         del username
-        return "Error: DEPRECATED option. Update your CASSH >= 1.5.0"
+        return response_render(
+            'Error: DEPRECATED option. Update your CASSH >= 1.5.0',
+            http_code='299 Deprecated')
 
     def POST(self, username):
         """
@@ -458,28 +192,28 @@ class Admin():
         # LDAP authentication
         is_admin_auth, message = ldap_authentification(admin=True)
         if not is_admin_auth:
-            ctx.status = '401 Unauthorized'
-            return message
+            return response_render(message, http_code='401 Unauthorized')
 
         payload = data2map()
 
-        if payload.has_key('revoke'):
+        if 'revoke' in payload:
             do_revoke = payload['revoke'].lower() == 'true'
         else:
             do_revoke = False
-        if payload.has_key('status'):
+        if 'status' in payload:
             do_status = payload['status'].lower() == 'true'
         else:
             do_status = False
 
-        pg_conn, message = pg_connection()
+        pg_conn, message = TOOLS.pg_connection()
         if pg_conn is None:
-            ctx.status = '503 Service Unavailable'
-            return message
+            return response_render(message, http_code='503 Service Unavailable')
         cur = pg_conn.cursor()
 
         if username == 'all' and do_status:
-            return list_keys()
+            return response_render(
+                TOOLS.list_keys(),
+                content_type='application/json')
 
         # Search if key already exists
         cur.execute('SELECT * FROM USERS WHERE NAME=(%s)', (username,))
@@ -488,17 +222,19 @@ class Admin():
         if user is None:
             cur.close()
             pg_conn.close()
-            message = "User '%s' does not exists." % username
+            message = 'User does not exists.'
         elif do_revoke:
             cur.execute('UPDATE USERS SET STATE=1 WHERE NAME=(%s)', (username,))
             pg_conn.commit()
             message = 'Revoke user=%s.' % username
-            cluster_updatekrl(username)
+            TOOLS.cluster_update_krl(username)
             cur.close()
             pg_conn.close()
         # Display status
         elif do_status:
-            return list_keys(username=username)
+            return response_render(
+                TOOLS.list_keys(username=username),
+                content_type='application/json')
         # If user is in PENDING state
         elif user[2] == 2:
             cur.execute('UPDATE USERS SET STATE=0 WHERE NAME=(%s)', (username,))
@@ -517,7 +253,7 @@ class Admin():
             cur.close()
             pg_conn.close()
             message = 'user=%s already active. Nothing done.' % username
-        return message
+        return response_render(message)
 
     def PATCH(self, username):
         """
@@ -528,13 +264,11 @@ class Admin():
         # LDAP authentication
         is_admin_auth, message = ldap_authentification(admin=True)
         if not is_admin_auth:
-            ctx.status = '401 Unauthorized'
-            return message
+            return response_render(message, http_code='401 Unauthorized')
 
-        pg_conn, message = pg_connection()
+        pg_conn, message = TOOLS.pg_connection()
         if pg_conn is None:
-            ctx.status = '503 Service Unavailable'
-            return message
+            return response_render(message, http_code='503 Service Unavailable')
         cur = pg_conn.cursor()
 
         payload = data2map()
@@ -543,29 +277,31 @@ class Admin():
             if key == 'expiry':
                 pattern = re_compile('^\\+([0-9]+)+[dh]$')
                 if pattern.match(value) is None:
-                    ctx.status = '400 Bad Request'
-                    return 'ERROR: Value %s is malformed. Should match pattern ^\\+([0-9]+)+[dh]$' \
-                        % value
+                    return response_render(
+                        'ERROR: Value %s is malformed. Should match pattern ^\\+([0-9]+)+[dh]$' \
+                        % value,
+                        http_code='400 Bad Request')
                 cur.execute('UPDATE USERS SET EXPIRY=(%s) WHERE NAME=(%s)', (value, username))
                 pg_conn.commit()
                 cur.close()
                 pg_conn.close()
-                return 'OK: %s=%s for %s' % (key, value, username)
+                return response_render('OK: %s=%s for %s' % (key, value, username))
             elif key == 'principals':
                 value = unquote_plus(value)
                 pattern = re_compile("^([a-zA-Z]+)$")
                 for principal in value.split(','):
                     if pattern.match(principal) is None:
-                        ctx.status = '400 Bad Request'
-                        return 'ERROR: Value %s is malformed. Should match pattern ^([a-zA-Z]+)$' \
-                            % principal
+                        return response_render(
+                            'ERROR: Value %s is malformed. Should match pattern ^([a-zA-Z]+)$' \
+                            % principal,
+                            http_code='400 Bad Request')
                 cur.execute('UPDATE USERS SET PRINCIPALS=(%s) WHERE NAME=(%s)', (value, username))
                 pg_conn.commit()
                 cur.close()
                 pg_conn.close()
-                return 'OK: %s=%s for %s' % (key, value, username)
+                return response_render('OK: %s=%s for %s' % (key, value, username))
 
-        return 'WARNING: No key found...'
+        return response_render('WARNING: No key found...')
 
     def DELETE(self, username):
         """
@@ -575,13 +311,11 @@ class Admin():
         # LDAP authentication
         is_admin_auth, message = ldap_authentification(admin=True)
         if not is_admin_auth:
-            ctx.status = '401 Unauthorized'
-            return message
+            return response_render(message, http_code='401 Unauthorized')
 
-        pg_conn, message = pg_connection()
+        pg_conn, message = TOOLS.pg_connection()
         if pg_conn is None:
-            ctx.status = '503 Service Unavailable'
-            return message
+            return response_render(message, http_code='503 Service Unavailable')
         cur = pg_conn.cursor()
 
         # Search if key already exists
@@ -589,7 +323,7 @@ class Admin():
         pg_conn.commit()
         cur.close()
         pg_conn.close()
-        return 'OK'
+        return response_render('OK')
 
 
 class Ca():
@@ -600,7 +334,9 @@ class Ca():
         """
         Return ca.
         """
-        return open(SERVER_OPTS['ca'] + '.pub', 'rb')
+        return response_render(
+            open(SERVER_OPTS['ca'] + '.pub', 'rb'),
+            content_type='application/octet-stream')
 
 class ClientStatus():
     """
@@ -614,18 +350,20 @@ class ClientStatus():
         # LDAP authentication
         is_auth, message = ldap_authentification()
         if not is_auth:
-            ctx.status = '401 Unauthorized'
-            return message
+            return response_render(message, http_code='401 Unauthorized')
 
         payload = data2map()
 
-        if payload.has_key('realname'):
+        if 'realname' in payload:
             realname = unquote_plus(payload['realname'])
         else:
-            ctx.status = '400 Bad Request'
-            return "Error: No realname option given."
+            return response_render(
+                'Error: No realname option given.',
+                http_code='400 Bad Request')
 
-        return list_keys(realname=realname)
+        return response_render(
+            TOOLS.list_keys(realname=realname),
+            content_type='application/json')
 
 class Client():
     """
@@ -636,8 +374,9 @@ class Client():
         DEPRECATED Status
         """
         del username
-        ctx.status = '299 Deprecated'
-        return "Error: DEPRECATED option. Update your CASSH >= 1.5.0"
+        return response_render(
+            'Error: DEPRECATED option. Update your CASSH >= 1.5.0',
+            http_code='299 Deprecated')
 
     def POST(self):
         """
@@ -652,8 +391,7 @@ class Client():
         # LDAP authentication
         is_auth, message = ldap_authentification()
         if not is_auth:
-            ctx.status = '401 Unauthorized'
-            return message
+            return response_render(message, http_code='401 Unauthorized')
 
         # Check if user is an admin and want to force signature when db fail
         force_sign = False
@@ -664,55 +402,59 @@ class Client():
         payload = data2map()
 
         if is_admin_auth and SERVER_OPTS['admin_db_failover'] \
-            and payload.has_key('admin_force') and payload['admin_force'].lower() == 'true':
+            and 'admin_force' in payload and payload['admin_force'].lower() == 'true':
             force_sign = True
 
         # Get username
-        if payload.has_key('username'):
+        if 'username' in payload:
             username = payload['username']
         else:
-            ctx.status = '400 Bad Request'
-            return "Error: No username option given. Update your CASSH >= 1.3.0"
+            return response_render(
+                'Error: No username option given. Update your CASSH >= 1.3.0',
+                http_code='400 Bad Request')
         username_pattern = re_compile("^([a-z]+)$")
         if username_pattern.match(username) is None or username == 'all':
-            ctx.status = '400 Bad Request'
-            return "Error: Username %s doesn't match pattern %s" \
-                % (username, username_pattern.pattern)
+            return response_render(
+                "Error: Username doesn't match pattern %s" \
+                % username_pattern.pattern,
+                http_code='400 Bad Request')
 
         # Get realname
-        if payload.has_key('realname'):
+        if 'realname' in payload:
             realname = unquote_plus(payload['realname'])
         else:
-            ctx.status = '400 Bad Request'
-            return "Error: No realname option given."
+            return response_render(
+                'Error: No realname option given.',
+                http_code='400 Bad Request')
 
         # Get public key
-        if payload.has_key('pubkey'):
+        if 'pubkey' in payload:
             pubkey = unquote_custom(payload['pubkey'])
         else:
-            ctx.status = '400 Bad Request'
-            return "Error: No pubkey given."
+            return response_render(
+                'Error: No pubkey given.',
+                http_code='400 Bad Request')
         tmp_pubkey = NamedTemporaryFile(delete=False)
-        tmp_pubkey.write(pubkey)
+        tmp_pubkey.write(bytes(pubkey, 'utf-8'))
         tmp_pubkey.close()
 
         pubkey_fingerprint = get_fingerprint(tmp_pubkey.name)
         if pubkey_fingerprint == 'Unknown':
             remove(tmp_pubkey.name)
-            ctx.status = '422 Unprocessable Entity'
-            return 'Error : Public key unprocessable'
+            return response_render(
+                'Error : Public key unprocessable',
+                http_code='422 Unprocessable Entity')
 
-        pg_conn, message = pg_connection()
+        pg_conn, message = TOOLS.pg_connection()
         # Admin force signature case
         if pg_conn is None and force_sign:
-            cert_contents = sign_key(tmp_pubkey.name, username, '+12h', username)
+            cert_contents = TOOLS.sign_key(tmp_pubkey.name, username, '+12h', username)
             remove(tmp_pubkey.name)
-            return cert_contents
+            return response_render(cert_contents, content_type='application/octet-stream')
         # Else, if db is down it fails.
         elif pg_conn is None:
             remove(tmp_pubkey.name)
-            ctx.status = '503 Service Unavailable'
-            return message
+            return response_render(message, http_code='503 Service Unavailable')
         cur = pg_conn.cursor()
 
         # Search if key already exists
@@ -722,14 +464,17 @@ class Client():
             cur.close()
             pg_conn.close()
             remove(tmp_pubkey.name)
-            return 'Error : User or Key absent, add your key again.'
+            return response_render(
+                'Error : User or Key absent, add your key again.',
+                http_code='400 Bad Request')
 
         if username != user[0] or realname != user[1]:
             cur.close()
             pg_conn.close()
             remove(tmp_pubkey.name)
-            ctx.status = '401 Unauthorized'
-            return 'Error : (username, realname) couple mismatch.'
+            return response_render(
+                'Error : (username, realname) couple mismatch.',
+                http_code='401 Unauthorized')
 
         status = user[2]
         expiry = user[6]
@@ -739,15 +484,17 @@ class Client():
             cur.close()
             pg_conn.close()
             remove(tmp_pubkey.name)
-            return "Status: %s" % STATES[user[2]]
+            return response_render("Status: %s" % STATES[user[2]])
 
-        cert_contents = sign_key(tmp_pubkey.name, username, expiry, principals, db_cursor=cur)
+        cert_contents = TOOLS.sign_key(tmp_pubkey.name, username, expiry, principals, db_cursor=cur)
 
         remove(tmp_pubkey.name)
         pg_conn.commit()
         cur.close()
         pg_conn.close()
-        return cert_contents
+        return response_render(
+            cert_contents,
+            content_type='application/octet-stream')
 
     def PUT(self):
         """
@@ -759,50 +506,60 @@ class Client():
         # LDAP authentication
         is_auth, message = ldap_authentification()
         if not is_auth:
-            ctx.status = '401 Unauthorized'
-            return message
+            return response_render(message, http_code='401 Unauthorized')
 
         payload = data2map()
 
-        if payload.has_key('username'):
+        if 'username' in payload:
             username = payload['username']
         else:
-            ctx.status = '400 Bad Request'
-            return "Error: No username option given."
+            return response_render(
+                'Error: No username option given.',
+                http_code='400 Bad Request')
 
         username_pattern = re_compile("^([a-z]+)$")
         if username_pattern.match(username) is None or username == 'all':
-            ctx.status = '400 Bad Request'
-            return "Error: Username %s doesn't match pattern %s" \
-                % (username, username_pattern.pattern)
+            return response_render(
+                "Error: Username doesn't match pattern %s" \
+                % username_pattern.pattern,
+                http_code='400 Bad Request')
 
-        if payload.has_key('realname'):
+
+        if 'realname' in payload:
             realname = unquote_plus(payload['realname'])
         else:
-            ctx.status = '400 Bad Request'
-            return "Error: No realname option given."
+            return response_render(
+                'Error: No realname option given.',
+                http_code='400 Bad Request')
+
+        realname_pattern = re_compile("^([a-zA-Z0-9\.]+@[a-zA-Z0-9\.]+)$")
+        if realname_pattern.match(realname) is None:
+            return response_render(
+                "Error: Realname doesn't match pattern",
+                http_code='400 Bad Request')
 
         # Get public key
-        if payload.has_key('pubkey'):
+        if 'pubkey' in payload:
             pubkey = unquote_custom(payload['pubkey'])
         else:
-            ctx.status = '400 Bad Request'
-            return 'Error: No pubkey given.'
+            return response_render(
+                'Error: No pubkey given.',
+                http_code='400 Bad Request')
         tmp_pubkey = NamedTemporaryFile(delete=False)
-        tmp_pubkey.write(pubkey)
+        tmp_pubkey.write(bytes(pubkey, 'utf-8'))
         tmp_pubkey.close()
 
         pubkey_fingerprint = get_fingerprint(tmp_pubkey.name)
         if pubkey_fingerprint == 'Unknown':
             remove(tmp_pubkey.name)
-            ctx.status = '422 Unprocessable Entity'
-            return 'Error : Public key unprocessable'
+            return response_render(
+                'Error : Public key unprocessable',
+                http_code='422 Unprocessable Entity')
 
-        pg_conn, message = pg_connection()
+        pg_conn, message = TOOLS.pg_connection()
         if pg_conn is None:
             remove(tmp_pubkey.name)
-            ctx.status = '503 Service Unavailable'
-            return message
+            return response_render(message, http_code='503 Service Unavailable')
         cur = pg_conn.cursor()
 
         # Search if key already exists
@@ -818,8 +575,9 @@ class Client():
             cur.close()
             pg_conn.close()
             remove(tmp_pubkey.name)
-            ctx.status = '201 Created'
-            return 'Create user=%s. Pending request.' % username
+            return response_render(
+                'Create user=%s. Pending request.' % username,
+                http_code='201 Created')
         else:
             # Check if realname is the same
             cur.execute('SELECT * FROM USERS WHERE NAME=(%s) AND REALNAME=lower((%s))', \
@@ -829,8 +587,9 @@ class Client():
                 cur.close()
                 pg_conn.close()
                 remove(tmp_pubkey.name)
-                ctx.status = '401 Unauthorized'
-                return 'Error : (username, realname) couple mismatch.'
+                return response_render(
+                    'Error : (username, realname) couple mismatch.',
+                    http_code='401 Unauthorized')
             # Update entry into database
             cur.execute('UPDATE USERS SET SSH_KEY=(%s), SSH_KEY_HASH=(%s), STATE=2, EXPIRATION=0 \
                 WHERE NAME=(%s)', (pubkey, pubkey_fingerprint, username))
@@ -838,7 +597,7 @@ class Client():
             cur.close()
             pg_conn.close()
             remove(tmp_pubkey.name)
-            return 'Update user=%s. Pending request.' % username
+            return response_render('Update user=%s. Pending request.' % username)
 
 
 class ClusterUpdateKRL():
@@ -854,26 +613,27 @@ class ClusterUpdateKRL():
         payload = data2map()
 
         # Check clustersecret
-        if payload.has_key('clustersecret'):
+        if 'clustersecret' in payload:
             if payload['clustersecret'] != SERVER_OPTS['clustersecret']:
-                ctx.status = '401 Unauthorized'
-                return 'Unauthorized'
+                return response_render(
+                    'Unauthorized',
+                    http_code='401 Unauthorized')
         else:
-            ctx.status = '401 Unauthorized'
-            return 'Unauthorized'
+            return response_render(
+                'Unauthorized',
+                http_code='401 Unauthorized')
 
         # Get username
-        if payload.has_key('username'):
+        if 'username' in payload:
             username = payload['username']
         else:
             # It means I need to update my krl to the latest version
-            cluster_last_krl()
-            return 'Update Complete'
+            TOOLS.cluster_last_krl()
+            return response_render('Update Complete')
 
-        pg_conn, message = pg_connection()
+        pg_conn, message = TOOLS.pg_connection()
         if pg_conn is None:
-            ctx.status = '503 Service Unavailable'
-            return message
+            return response_render(message, http_code='503 Service Unavailable')
         cur = pg_conn.cursor()
 
         message = 'Revoke user=%s.' % username
@@ -882,14 +642,14 @@ class ClusterUpdateKRL():
         cur.execute('SELECT SSH_KEY FROM USERS WHERE NAME=(%s)', (username,))
         pubkey = cur.fetchone()[0]
         tmp_pubkey = NamedTemporaryFile(delete=False)
-        tmp_pubkey.write(pubkey)
+        tmp_pubkey.write(bytes(pubkey, 'utf-8'))
         tmp_pubkey.close()
         ca_ssh.update_krl(tmp_pubkey.name)
         cur.close()
         pg_conn.close()
         remove(tmp_pubkey.name)
 
-        return 'Revoke Complete'
+        return response_render('Revoke Complete')
 
 class ClusterStatus():
     """
@@ -900,15 +660,14 @@ class ClusterStatus():
         /cluster/status
         """
         message = dict()
-
-        alive_nodes, dead_nodes = cluster_alived()
-
+        alive_nodes, dead_nodes = TOOLS.cluster_alived()
         for node in alive_nodes:
             message.update({node: {'status': 'OK'}})
         for node in dead_nodes:
             message.update({node: {'status': 'KO'}})
-
-        return dumps(message)
+        return response_render(
+            dumps(message),
+            content_type='application/json')
 
 
 class Health():
@@ -922,7 +681,9 @@ class Health():
         health = {}
         health['name'] = 'cassh'
         health['version'] = VERSION
-        return dumps(health, indent=4, sort_keys=True)
+        return response_render(
+            dumps(health, indent=4, sort_keys=True),
+            content_type='application/json')
 
 
 class Krl():
@@ -933,7 +694,9 @@ class Krl():
         """
         Return krl.
         """
-        return open(SERVER_OPTS['krl'], 'rb')
+        return response_render(
+            open(SERVER_OPTS['krl'], 'rb'),
+            content_type='application/octet-stream')
 
 
 class Ping():
@@ -944,7 +707,7 @@ class Ping():
         """
         Return a pong
         """
-        return 'pong'
+        return response_render('pong')
 
 
 class TestAuth():
@@ -958,9 +721,8 @@ class TestAuth():
         # LDAP authentication
         is_auth, message = ldap_authentification()
         if not is_auth:
-            ctx.status = '401 Unauthorized'
-            return message
-        return 'OK'
+            return response_render(message, http_code='401 Unauthorized')
+        return response_render('OK')
 
 
 class MyApplication(application):
@@ -980,6 +742,8 @@ if __name__ == "__main__":
         print('LDAP: %s' % SERVER_OPTS['ldap'])
         print('Admin DB Failover: %s' % SERVER_OPTS['admin_db_failover'])
     APP = MyApplication(URLS, globals())
-    config.debug = False
-    cluster_last_krl()
+    config.debug = SERVER_OPTS['debug']
+    if SERVER_OPTS['debug']:
+        print('Debug mode on')
+    TOOLS.cluster_last_krl()
     APP.run()
