+++ src/server/sql/revocation.sql
@@ -1,0 +1,8 @@
+-------------------------
+-- CASSH Revocation table
+-------------------------
+CREATE TABLE REVOCATION
+(
+    SSH_KEY            TEXT  PRIMARY KEY  NOT NULL,
+    REVOCATION_DATE    INT                NOT NULL
+)
