+++ tests/docker-compose.yml
@@ -30,7 +30,7 @@ services:
       POSTGRES_DB: cassh
     # Init the DB
     volumes:
-      - ../src/server/sql/model.sql:/docker-entrypoint-initdb.d/cassh_model.sql
+      - ../src/server/sql/users.sql:/docker-entrypoint-initdb.d/cassh_users.sql
 
   #TODO: LDAP
